+++ src/components/Cards/TasksCard.jsx
@@ -1,17 +1,19 @@
 import React from "react";
 import PropTypes from "prop-types";
-import {
-  withStyles,
-  Card,
-  CardContent,
-  CardHeader,
-  Typography,
-  Tabs,
-  Tab
-} from "material-ui";
-import { BugReport, Code, Cloud } from "@material-ui/icons";
-
-import { Tasks } from "components";
+// @material-ui/core components
+import withStyles from "@material-ui/core/styles/withStyles";
+import Card from "@material-ui/core/Card";
+import CardContent from "@material-ui/core/CardContent";
+import CardHeader from "@material-ui/core/CardHeader";
+import Typography from "@material-ui/core/Typography";
+import Tabs from "@material-ui/core/Tabs";
+import Tab from "@material-ui/core/Tab";
+// @material-ui/icons
+import BugReport from "@material-ui/icons/BugReport";
+import Code from "@material-ui/icons/Code";
+import Cloud from "@material-ui/icons/Cloud";
+// core components
+import Tasks from "components/Tasks/Tasks.jsx";
 
 import { bugs, website, server } from "variables/general";