+++ src/components/Footer/Footer.jsx
@@ -1,7 +1,10 @@
 import React from "react";
 import PropTypes from "prop-types";
-import { List, ListItem, withStyles } from "material-ui";
-
+// @material-ui/core components
+import withStyles from "@material-ui/core/styles/withStyles";
+import ListItem from "@material-ui/core/ListItem";
+import List from "@material-ui/core/List";
+// core components
 import footerStyle from "assets/jss/material-dashboard-react/footerStyle";
 
 function Footer({ ...props }) {