+++ src/components/Sidebar/Sidebar.jsx
@@ -1,18 +1,17 @@
 import React from "react";
+import classNames from "classnames";
 import PropTypes from "prop-types";
 import { NavLink } from "react-router-dom";
-import cx from "classnames";
-import {
-  withStyles,
-  Drawer,
-  Hidden,
-  List,
-  ListItem,
-  ListItemIcon,
-  ListItemText
-} from "material-ui";
-
-import { HeaderLinks } from "components";
+// @material-ui/core components
+import withStyles from "@material-ui/core/styles/withStyles";
+import Drawer from "@material-ui/core/Drawer";
+import Hidden from "@material-ui/core/Hidden";
+import List from "@material-ui/core/List";
+import ListItem from "@material-ui/core/ListItem";
+import ListItemIcon from "@material-ui/core/ListItemIcon";
+import ListItemText from "@material-ui/core/ListItemText";
+// core components
+import HeaderLinks from "components/Header/HeaderLinks.jsx";
 
 import sidebarStyle from "assets/jss/material-dashboard-react/sidebarStyle.jsx";
 
@@ -26,10 +25,10 @@ const Sidebar = ({ ...props }) => {
     <List className={classes.list}>
       {routes.map((prop, key) => {
         if (prop.redirect) return null;
-        const listItemClasses = cx({
+        const listItemClasses = classNames({
           [" " + classes[color]]: activeRoute(prop.path)
         });
-        const whiteFontClasses = cx({
+        const whiteFontClasses = classNames({
           [" " + classes.whiteFont]: activeRoute(prop.path)
         });
         return (