+++ src/components/Snackbar/Snackbar.jsx
@@ -1,15 +1,19 @@
 import React from "react";
-import { withStyles, Snackbar as Snack, IconButton } from "material-ui";
-import { Close } from "@material-ui/icons";
+import classNames from "classnames";
 import PropTypes from "prop-types";
-import cx from "classnames";
-
+// @material-ui/core components
+import withStyles from "@material-ui/core/styles/withStyles";
+import Snack from "@material-ui/core/Snackbar";
+import IconButton from "@material-ui/core/IconButton";
+// @material-ui/icons
+import Close from "@material-ui/icons/Close";
+// core components
 import snackbarContentStyle from "assets/jss/material-dashboard-react/snackbarContentStyle.jsx";
 
 function Snackbar({ ...props }) {
   const { classes, message, color, close, icon, place, open } = props;
   var action = [];
-  const messageClasses = cx({
+  const messageClasses = classNames({
     [classes.iconMessage]: icon !== undefined
   });
   if (close !== undefined) {