+++ src/components/Table/Table.jsx
@@ -1,15 +1,13 @@
 import React from "react";
-import {
-  withStyles,
-  Table,
-  TableHead,
-  TableRow,
-  TableBody,
-  TableCell
-} from "material-ui";
-
 import PropTypes from "prop-types";
-
+// @material-ui/core components
+import withStyles from "@material-ui/core/styles/withStyles";
+import Table from "@material-ui/core/Table";
+import TableHead from "@material-ui/core/TableHead";
+import TableRow from "@material-ui/core/TableRow";
+import TableBody from "@material-ui/core/TableBody";
+import TableCell from "@material-ui/core/TableCell";
+// core components
 import tableStyle from "assets/jss/material-dashboard-react/tableStyle";
 
 function CustomTable({ ...props }) {