+++ src/components/Tasks/Tasks.jsx
@@ -1,18 +1,19 @@
 import React from "react";
-import {
-  withStyles,
-  Checkbox,
-  IconButton,
-  Table,
-  TableBody,
-  TableCell,
-  TableRow,
-  Tooltip
-} from "material-ui";
-import { Edit, Close, Check } from "@material-ui/icons";
-
 import PropTypes from "prop-types";
-
+// @material-ui/core components
+import withStyles from "@material-ui/core/styles/withStyles";
+import Checkbox from "@material-ui/core/Checkbox";
+import Tooltip from "@material-ui/core/Tooltip";
+import IconButton from "@material-ui/core/IconButton";
+import Table from "@material-ui/core/Table";
+import TableRow from "@material-ui/core/TableRow";
+import TableBody from "@material-ui/core/TableBody";
+import TableCell from "@material-ui/core/TableCell";
+// @material-ui/icons
+import Edit from "@material-ui/icons/Edit";
+import Close from "@material-ui/icons/Close";
+import Check from "@material-ui/icons/Check";
+// core components
 import tasksStyle from "assets/jss/material-dashboard-react/tasksStyle.jsx";
 
 class Tasks extends React.Component {