+++ src/components/Typography/Quote.jsx
@@ -1,7 +1,8 @@
 import React from "react";
-import { withStyles } from "material-ui";
 import PropTypes from "prop-types";
-
+// @material-ui/core components
+import withStyles from "@material-ui/core/styles/withStyles";
+// core components
 import typographyStyle from "assets/jss/material-dashboard-react/typographyStyle.jsx";
 
 function Quote({ ...props }) {