+++ CHANGELOG.md
@@ -1,3 +1,46 @@
+## [1.3.0] 2018-04-16
+### Bug Fixing
+- Changed import from `material-ui` to `@material-ui/core`
+- Droped `src/components/index.js`, and changed all the imports to separate ones
+- Renamed `ItemGrid` to `GridItem`
+### Major style changes
+### Dropped components
+- `src/components/CustomButtons/IconButton.jsx` (instead use `src/components/CustomButtons/Button.jsx` with `justIcon` prop)
+- Some components from `src/components/Typography` (instead of these we've added some css to style the normal HTML tags)
+  - `src/components/Typography/A.jsx`
+  - `src/components/Typography/P.jsx`
+  - `src/components/Typography/Small.jsx`
+- All the cards from `src/components/Cards`
+  - `src/components/Cards/ChartCard.jsx`
+  - `src/components/Cards/ProfileCard.jsx`
+  - `src/components/Cards/RegularCard.jsx`
+  - `src/components/Cards/StatsCard.jsx`
+  - `src/components/Cards/TasksCard.jsx`
+### Added components
+- New card components (`src/components/Card/*`) instead of `src/components/Cards/*`
+  -
+  -
+  -
+  -
+  -
+### Deleted dependencies
+- `material-ui@1.0.0-beta.41`
+### Added dependencies
+- `@material-ui/core@1.2.1` (instead of `material-ui@1.0.0-beta.41`)
+- `@types/googlemaps@3.30.9` to stop the warning: **npm WARN react-google-maps@9.4.5 requires a peer of @types/googlemaps@^3.0.0 but none is installed. You must install peer dependencies yourself.**
+- `@types/markerclustererplus@2.1.33` to stop the warning: **npm WARN react-google-maps@9.4.5 requires a peer of @types/markerclustererplus@^2.1.29 but none is installed. You must install peer dependencies yourself.**
+- `ajv@6.5.1` to stop the warning: **npm WARN ajv-keywords@3.2.0 requires a peer of ajv@^6.0.0 but none is installed. You must install peer dependencies yourself.**
+### Update dependencies
+- `@material-ui/icons@1.0.0-beta.42` to `@material-ui/icons@1.1.0`
+- `classnames@2.2.5` to `classnames@2.2.6`
+- `npm-run-all@4.1.2` to `npm-run-all@4.1.3`
+- `perfect-scrollbar@1.3.0` to `perfect-scrollbar@1.4.0`
+- `react@16.2.0` to `react@16.4.0`
+- `react-dom@16.2.0` to `react-dom@16.4.0`
+- `react-router-dom@4.2.2` to `react-router-dom@4.3.1`
+- `react-scripts@1.0.17` to `react-scripts@1.1.4`
+- `react-swipeable-views@0.12.12` to `react-swipeable-views@0.12.13`
+
 ## [1.2.0] 2018-04-16
 ### Bug Fixing
 - Renamed `src/containers/App/App.jsx` to `src/layouts/Dashboard/Dashboard.jsx`